Refactored and moved lifecycle hooks to graphql package (out of JMS data store).
