+++ elide-datastore/elide-datastore-jms/src/main/java/com/yahoo/elide/datastores/jms/JMSDataStore.java
@@ -6,24 +6,14 @@
 
 package com.yahoo.elide.datastores.jms;
 
-import com.yahoo.elide.annotation.LifeCycleHookBinding;
-import com.yahoo.elide.annotation.Subscription;
-import com.yahoo.elide.annotation.SubscriptionField;
 import com.yahoo.elide.core.datastore.DataStore;
 import com.yahoo.elide.core.datastore.DataStoreTransaction;
 import com.yahoo.elide.core.dictionary.EntityDictionary;
-import com.yahoo.elide.core.type.ClassType;
 import com.yahoo.elide.core.type.Type;
-import com.yahoo.elide.core.utils.ClassScanner;
-import com.yahoo.elide.datastores.jms.hooks.NotifyTopicLifeCycleHook;
-
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.google.common.base.Preconditions;
-
 import lombok.Builder;
 
 import java.util.Set;
-import java.util.stream.Collectors;
 import javax.jms.ConnectionFactory;
 import javax.jms.JMSContext;
 
@@ -62,87 +52,11 @@ public class JMSDataStore implements DataStore {
         this.timeoutInMs = timeoutInMs;
     }
 
-    /**
-     * Constructor.
-     * @param scanner Used to scan for subscription classes.
-     * @param connectionFactory The JMS connection factory.
-     * @param dictionary The entity dictionary.
-     * @param mapper Object mapper for serializing/deserializing elide models to JMS topics.
-     * @param timeoutInMs request timeout in milliseconds.  0 means immediate.  -1 means no timeout.
-     */
-    public JMSDataStore(
-            ClassScanner scanner,
-            ConnectionFactory connectionFactory,
-            EntityDictionary dictionary,
-            ObjectMapper mapper,
-            long timeoutInMs
-    ) {
-        this(
-                scanner.getAnnotatedClasses(Subscription.class).stream()
-                        .map(ClassType::of)
-                        .collect(Collectors.toSet()),
-                connectionFactory,
-                dictionary,
-                mapper,
-                timeoutInMs
-        );
-    }
-
     @Override
     public void populateEntityDictionary(EntityDictionary dictionary) {
         for (Type<?> model : models) {
             dictionary.bindEntity(model);
-            Subscription subscription = model.getAnnotation(Subscription.class);
-            Preconditions.checkNotNull(subscription);
-
-            Subscription.Operation[] operations = subscription.operations();
-
-            for (Subscription.Operation operation : operations) {
-                switch (operation) {
-                    case UPDATE: {
-                        addUpdateHooks(model, dictionary);
-                        break;
-                    }
-                    case DELETE: {
-                        dictionary.bindTrigger(
-                                model,
-                                LifeCycleHookBinding.Operation.DELETE,
-                                LifeCycleHookBinding.TransactionPhase.POSTCOMMIT,
-                                new NotifyTopicLifeCycleHook(connectionFactory, mapper),
-                                false
-                        );
-                        break;
-                    }
-                    case CREATE: {
-                        dictionary.bindTrigger(
-                                model,
-                                LifeCycleHookBinding.Operation.CREATE,
-                                LifeCycleHookBinding.TransactionPhase.POSTCOMMIT,
-                                new NotifyTopicLifeCycleHook(connectionFactory, mapper),
-                                false
-                        );
-                        break;
-                    }
-                }
-            }
         }
-    }
-
-    protected void addUpdateHooks(Type<?> model, EntityDictionary dictionary) {
-        dictionary.getAllFields(model).stream().forEach(fieldName -> {
-            SubscriptionField subscriptionField =
-                    dictionary.getAttributeOrRelationAnnotation(model, SubscriptionField.class, fieldName);
-
-            if (subscriptionField != null) {
-                dictionary.bindTrigger(
-                        model,
-                        fieldName,
-                        LifeCycleHookBinding.Operation.UPDATE,
-                        LifeCycleHookBinding.TransactionPhase.POSTCOMMIT,
-                        new NotifyTopicLifeCycleHook(connectionFactory, mapper)
-                );
-            }
-        });
     }
 
     @Override
