+++ elide-graphql/src/main/java/com/yahoo/elide/graphql/GraphQLNameUtils.java
@@ -6,9 +6,9 @@
 
 package com.yahoo.elide.graphql;
 
-import com.yahoo.elide.annotation.Subscription;
 import com.yahoo.elide.core.dictionary.EntityDictionary;
 import com.yahoo.elide.core.type.Type;
+import com.yahoo.elide.graphql.subscriptions.annotations.Subscription;
 import org.apache.commons.lang3.StringUtils;
 
 public class GraphQLNameUtils {
