+++ elide-graphql/src/main/java/com/yahoo/elide/graphql/subscriptions/SubscriptionModelBuilder.java
@@ -8,8 +8,6 @@ package com.yahoo.elide.graphql.subscriptions;
 import static graphql.schema.GraphQLArgument.newArgument;
 import static graphql.schema.GraphQLFieldDefinition.newFieldDefinition;
 import static graphql.schema.GraphQLObjectType.newObject;
-import com.yahoo.elide.annotation.Subscription;
-import com.yahoo.elide.annotation.SubscriptionField;
 import com.yahoo.elide.core.dictionary.EntityDictionary;
 import com.yahoo.elide.core.dictionary.RelationshipType;
 import com.yahoo.elide.core.type.Type;
@@ -17,6 +15,8 @@ import com.yahoo.elide.graphql.GraphQLConversionUtils;
 import com.yahoo.elide.graphql.GraphQLNameUtils;
 import com.yahoo.elide.graphql.GraphQLScalars;
 import com.yahoo.elide.graphql.NonEntityDictionary;
+import com.yahoo.elide.graphql.subscriptions.annotations.Subscription;
+import com.yahoo.elide.graphql.subscriptions.annotations.SubscriptionField;
 import com.google.common.collect.Sets;
 import graphql.Scalars;
 import graphql.schema.DataFetcher;
