+++ elide-graphql/src/main/java/com/yahoo/elide/graphql/subscriptions/annotations/SubscriptionField.java
@@ -3,7 +3,7 @@
  * Licensed under the Apache License, Version 2.0
  * See LICENSE file in project root for terms.
  */
-package com.yahoo.elide.annotation;
+package com.yahoo.elide.graphql.subscriptions.annotations;
 
 import static java.lang.annotation.ElementType.FIELD;
 import static java.lang.annotation.ElementType.METHOD;
