+++ elide-graphql/src/main/java/com/yahoo/elide/graphql/subscriptions/containers/SubscriptionNodeContainer.java
@@ -9,6 +9,7 @@ import com.yahoo.elide.core.PersistentResource;
 import com.yahoo.elide.core.request.Relationship;
 import com.yahoo.elide.graphql.Environment;
 import com.yahoo.elide.graphql.containers.NodeContainer;
+
 import java.util.LinkedHashSet;
 import java.util.List;
 import java.util.Set;
