+++ elide-graphql/src/main/java/com/yahoo/elide/graphql/subscriptions/hooks/TopicType.java
@@ -4,7 +4,7 @@
  * See LICENSE file in project root for terms.
  */
 
-package com.yahoo.elide.datastores.jms;
+package com.yahoo.elide.graphql.subscriptions.hooks;
 
 import static com.yahoo.elide.annotation.LifeCycleHookBinding.Operation.CREATE;
 import static com.yahoo.elide.annotation.LifeCycleHookBinding.Operation.DELETE;
