+++ elide-graphql/src/main/java/com/yahoo/elide/graphql/subscriptions/websocket/ConnectionInfo.java
@@ -7,7 +7,6 @@
 package com.yahoo.elide.graphql.subscriptions.websocket;
 
 import com.yahoo.elide.core.security.User;
-
 import lombok.Builder;
 import lombok.Value;
