+++ elide-graphql/src/test/java/com/yahoo/elide/graphql/subscriptions/hooks/NotifyTopicLifeCycleHookTest.java
@@ -4,7 +4,7 @@
  * See LICENSE file in project root for terms.
  */
 
-package com.yahoo.elide.datastores.jms.hooks;
+package com.yahoo.elide.graphql.subscriptions.hooks;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.mockito.ArgumentMatchers.any;
